social-media/weibo-cli/pyproject.toml
[project]
name = "weibo-cli"
version = "0.1.0"
description = "A CLI for Weibo (微博) — search, browse hot topics, read timelines, and explore user profiles from the terminal"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "xiesg", email = "xiesg@example.com" }
]
requires-python = ">=3.10"
keywords = ["weibo", "sina", "微博", "cli", "social-media"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Office/Business",
]
dependencies = [
    "click>=8.0",
    "httpx>=0.27",
    "rich>=13.0",
    "browser-cookie3>=0.19",
    "qrcode>=7.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-requests>=2.31",
]

[project.scripts]
weibo = "weibo_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
